DBConfig: - extracted the db-config-file-logic into its own class - added a mechanism to checkpoint db-config-files (this won't fix the replication load-mechanism)
Replication: expanded the interface with operations to control babudb remotely to support the new introduced declareToMaster-function of the ReplicationManager. This operation is necessary to support master-slave-failover with fatlease. The Operations are not implemented jet.

Changed the behavior of the stop-restart-functionality of the replication to match the new requirements necessary for the declareToMaster-function.

ReplicationManager: introduction of the declareToMaster-function, which will make babuDB load from the latest available slave and declare itself to master while remotely switching all other participants into slaves

git-svn-id: http://babudb.googlecode.com/svn/trunk@141 d4c7dd96-b7b2-11dd-893d-d98698f66bbf